requirements.txt
openai>=1.0.0
tiktoken>=0.5.0
yt-dlp>=2023.12.30
openai-whisper>=20231117
trafilatura>=1.6.1
# marker-pdf provides PDF text extraction
marker-pdf>=0.1.0
# pypdfium2 can be installed for fallback extraction
# marker-ocr-pdf>=0.1.0  # このパッケージは利用できないためコメントアウト
language-tool-python>=2.7.1
sacrebleu>=2.3.1
pytest>=7.4.0
ruff>=0.1.0
mypy>=1.7.0
click>=8.0
streamlit>=1.29.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
numpy>=1.24.0
torch>=2.1.0
transformers>=4.36.0
PyYAML>=6.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
langdetect>=1.0.9
fugashi>=1.3.0
patch-ng>=1.18.1
